+++ src/main/java/cs319/cards/model/Card.java
@@ -2,7 +2,7 @@ package cs319.cards.model;
 
 public interface Card {
 
-    short getCardId();
+    int getCardId();
 
     String getCardMessage();
 }
